refactor: colocate updateFiles callback with DropZone component and modify types
