+++ src/components/internal/FormFields/FileInput/constants.ts
@@ -1,2 +1,2 @@
 export const DEFAULT_FILE_TYPES = ['application/pdf', 'image/jpeg', 'image/jpg', 'image/png'] as const;
-export const DEFAULT_MAX_SIZE = 2097152; // 2MB
+export const DEFAULT_MAX_FILE_SIZE = 2097152; // 2MB
